oom: remove oom_disable_count
This removes mm->oom_disable_count entirely since it's unnecessary
and
currently buggy.  The counter was intended to be per-process but
it's
currently decremented in the exit path for each thread that exits,
causing
it to underflow.

The count was originally intended to prevent
oom killing threads that
share memory with threads that cannot be killed
since it doesn't lead to
future memory freeing.  The counter could be
fixed to represent all
threads sharing the same mm, but it's better to
remove the count since:

 - it is possible that the OOM_DISABLE thread
sharing memory with the
   victim is waiting on that thread to exit and
will actually cause
   future memory freeing, and

 - there is no
guarantee that a thread is disabled from oom killing just
   because
another thread sharing its mm is oom disabled.

Change-Id:
I2aa1be2a9291adee53af923e3c7ac670dad7b8c0
Signed-off-by: David Rientjes
<rientjes@google.com>
Reported-by: Oleg Nesterov
<oleg@redhat.com>
Reviewed-by: Oleg Nesterov <oleg@redhat.com>
Cc: Ying
Han <yinghan@google.com>
Cc: KOSAKI Motohiro
<kosaki.motohiro@jp.fujitsu.com>
Signed-off-by: Andrew Morton
<akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds
<torvalds@linux-foundation.org>